# docs/swagger.yaml
definitions:
  deliverymodels.AddProductInputs:
    properties:
      branid:
        type: integer
      description:
        type: string
      name:
        type: string
    required:
    - description
    - name
    type: object
  deliverymodels.AddProductsDeatils:
    properties:
      price:
        type: number
      productid:
        type: integer
      quantity:
        type: integer
      sizeid:
        type: integer
    required:
    - quantity
    type: object
...
  deliverymodels.AdminSignupInputs:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        type: string
      phonenumber:
        type: string
    type: object
  deliverymodels.AdmionLoginInputs:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  deliverymodels.Coupon:
    properties:
      couponCode:
        type: string
      discount:
        type: integer
      expirDate:
        type: string
      maximumdiscount:
        type: integer
    type: object
  deliverymodels.UpdateProductInput:
    properties:
      branid:
        type: integer
      description:
        type: string
      name:
        description: AdminId     uint   `json:"adminid" `
        type: string
      productid:
        type: integer
    required:
    - description
    - name
    type: object
  deliverymodels.UpdateProductsDeatils:
    properties:
      price:
        type: number
      productId:
        description: AdminId       uint `json:"adminId" `
        type: integer
      productItemId:
        type: integer
      quantity:
        type: integer
      sizeid:
        type: integer
    required:
    - quantity
    type: object
...
  deliverymodels.ViewProductDetail:
    properties:
      brand:
        type: string
      description:
        type: string
      name:
        type: string
      price:
        type: number
      sizes:
        items:
          type: string
        type: array
    type: object
...
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: Testing Swagger APIs.
  termsOfService: http://swagger.io/terms/
  title: 70_Off
  version: "1.0"
paths:
  /adminLogin:
    post:
      consumes:
      - application/json
      description: Admin Can Login With UserName And Password
      parameters:
      - description: Admin Data
        in: body
        name: admin
        required: true
        schema:
          $ref: '#/definitions/deliverymodels.AdmionLoginInputs'
      produces:
      - application/json
      responses:
        "200":
          description: OK
      summary: Admin Login
      tags:
      - Admin
  /adminSignup:
    post:
      consumes:
      - application/json
      description: Add new Admin
      parameters:
      - description: User Data
        in: body
        name: admin
        required: true
        schema:
          $ref: '#/definitions/deliverymodels.AdminSignupInputs'
      produces:
      - application/json
      responses: {}
      summary: Admin SignUp
      tags:
      - Admin
...
  /v1/admin/addCoupon:
    post:
      description: Add Coupon
      parameters:
      - description: Coupon Data
        in: body
        name: coupon
        required: true
        schema:
          $ref: '#/definitions/deliverymodels.Coupon'
      responses:
        "200":
          description: Success Message
          schema:
            type: string
      summary: Add Coupon
      tags:
      - Admin
  /v1/admin/addProduct:
    post:
      consumes:
      - application/json
      description: Admin Can Add Product
      parameters:
      - description: Product Data
        in: body
        name: admin
        required: true
        schema:
          $ref: '#/definitions/deliverymodels.AddProductInputs'
      produces:
      - application/json
      responses: {}
      summary: Add Product
      tags:
      - Admin
  /v1/admin/addProductDetails:
    post:
      consumes:
      - application/json
      description: Admin Can Add Product Details
      parameters:
      - description: Product Data
        in: body
        name: admin
        required: true
        schema:
          $ref: '#/definitions/deliverymodels.AddProductsDeatils'
      produces:
      - application/json
      responses: {}
      summary: Add Product Details
      tags:
      - Admin
...
  /v1/admin/delete/{productId}:
    put:
      consumes:
      - application/json
      description: Product Soft Delete
      parameters:
      - description: 'ProductID '
        in: path
        name: productID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: Soft Delete
      tags:
      - Admin
  /v1/admin/saleReport/{days}:
    get:
      consumes:
      - application/json
      description: It shows weekly,monthly,yearly and last n days order report
      parameters:
      - description: Days
        in: path
        name: days
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: Order Report
      tags:
      - Admin
  /v1/admin/updateProduct:
    put:
      consumes:
      - application/json
      description: Admin Can Edit Product
      parameters:
      - description: Product Data
        in: body
        name: admin
        required: true
        schema:
          $ref: '#/definitions/deliverymodels.UpdateProductInput'
      produces:
      - application/json
      responses: {}
      summary: Edit Product
      tags:
      - Admin
  /v1/admin/updateProductDetails:
    put:
      consumes:
      - application/json
      description: Admin Can Edit Product Details
      parameters:
      - description: Product Details
        in: body
        name: admin
        required: true
        schema:
          $ref: '#/definitions/deliverymodels.UpdateProductsDeatils'
      produces:
      - application/json
      responses: {}
      summary: Edit Product Details
      tags:
      - Admin
...
  /v1/admin/updatingReturnStatus/{orderItemID}:
    put:
      consumes:
      - application/json
      description: Admin Can Updates Return Status for Orders
      parameters:
      - description: OrderItemID
        in: path
        name: orderItemID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: Updating Return Status ByAdmin
      tags:
      - Admin
  /v1/admin/userBlockToggle/{userID}:
    put:
      consumes:
      - application/json
      description: Admin Can Block or Unblock User
      parameters:
      - description: 'UserID '
        in: path
        name: userID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: User BLock Toggle
      tags:
      - Admin
...
  /v1/productsPagination:
    get:
      description: List Products with Pagination
      parameters:
      - description: Page
        in: query
        name: page
        type: string
      - description: Limit
        in: query
        name: limit
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/deliverymodels.ViewProductDetail'
      summary: List Products Pagination
      tags:
      - user
...
securityDefinitions:
  JWT:
    in: header
    name: token
    type: apiKey
swagger: "2.0"